try the next host on connnect_timeout
libpq does not support connect_timeout when using the async api.

So when using the async connection API  with a multi-host connection
string libpq will not timeout connections to the first host in the
list and thus will not attempt to connect to any subsequent hosts
in the connection string list.

This fixes this by closing and reopenning the connection with a
reordered connection string when connection times out.

See discussion on the pgsql-hackers list discussing this "feature" of the api:
https://www.postgresql.org/message-id/flat/CA%2Bmi_8YyGKA9dWELu63e%3DKL2oN-%2BFe4uca4EtFfb6uQD4Up8pw%40mail.gmail.com